+++ src/ua/data_types/string.rs
@@ -10,11 +10,17 @@ crate::data_type!(String);
 // strings of `length` 0. It may also be `ptr::null()` for "invalid" strings. This is similar to how
 // OPC UA treats arrays (which also distinguishes between empty and invalid instances).
 impl String {
+    #[deprecated(note = "use `Self::as_bytes()` instead")]
+    #[must_use]
+    pub fn as_slice(&self) -> Option<&[u8]> {
+        self.as_bytes()
+    }
+
     /// Returns string contents as byte slice.
     ///
     /// This may return [`None`] when the string itself is invalid (as defined by OPC UA).
     #[must_use]
-    pub fn as_slice(&self) -> Option<&[u8]> {
+    pub fn as_bytes(&self) -> Option<&[u8]> {
         // Internally, `open62541` represents strings as `Byte` array and has the same special cases
         // as regular arrays, i.e. empty and invalid states.
         match ua::ArrayValue::from_ptr(self.0.data) {
@@ -33,7 +39,7 @@ impl String {
     /// not valid UTF-8.
     #[must_use]
     pub fn as_str(&self) -> Option<&str> {
-        self.as_slice().and_then(|slice| str::from_utf8(slice).ok())
+        self.as_bytes().and_then(|slice| str::from_utf8(slice).ok())
     }
 }